feat: Authenticator trait
