+++ miden-lib/asm/miden/contracts/auth/basic.masm
@@ -37,14 +37,14 @@ export.auth_tx_rpo_falcon512
     push.PUBLIC_KEY_SLOT exec.account::get_item
     # => [PUB_KEY, M]
 
+    # Update the nonce
+    push.1 exec.account::incr_nonce
+    # => []
+    
     # Verify the signature against the public key and the message. The procedure gets as
     # inputs the hash of the public key and the hash of the message via the operand
     # stack. The signature is provided via the advice stack. The signature is valid if and
     # only if the procedure returns.
     exec.rpo_falcon512::verify
-    # => []
-
-    # Update the nonce
-    push.1 exec.account::incr_nonce
     # => []
 end
