+++ mock/src/mock/host/mod.rs
@@ -71,10 +71,7 @@ impl Host for MockHost {
         process: &S,
         injector: AdviceInjector,
     ) -> Result<HostResponse, ExecutionError> {
-        match injector {
-            AdviceInjector::SigToStack => self.adv_provider.push_signature(process),
-            injector => self.adv_provider.set_advice(process, &injector),
-        }
+        self.adv_provider.set_advice(process, &injector)
     }
 
     fn on_event<S: ProcessState>(
